pyproject.toml
[project]
name = "project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "matplotlib>=3.10.8",
    "numba>=0.64.0",
    "numpy>=2.4.3",
    "pandas>=3.0.1",
    "pyyaml>=6.0.3",
    "scipy>=1.17.1",
    "seaborn>=0.13.2",
]
